Green Sense Project - NIQE Analyzer/requirements.txt
numpy
scipy
Pillow
streamlit
pandas
opencv-python
torch
transformers
matplotlib
